Update delete interview command

+++ src/main/java/seedu/application/logic/commands/InterviewDeleteCommand.java
@@ -1,7 +1,7 @@
 package seedu.application.logic.commands;
 
 import static java.util.Objects.requireNonNull;
-import static seedu.application.logic.parser.CliSyntax.PREFIX_INTERVIEW_DELETE;
+import static seedu.application.logic.parser.CliSyntax.PREFIX_JOB_SOURCE;
 
 import java.util.List;
 
@@ -21,7 +21,7 @@ public class InterviewDeleteCommand extends InterviewCommand {
 
     public static final String MESSAGE_USAGE = COMMAND_WORD + ": Deletes an interview from the job\n"
             + "Parameters: INDEX (of interview) and\n INDEX (of job application)\n"
-            + "Example: 1 " + PREFIX_INTERVIEW_DELETE + "2\n"
+            + "Example: 1 " + PREFIX_JOB_SOURCE + "2\n"
             + "deletes 1st interview from 2nd job";
 
     public static final String MESSAGE_SUCCESS = "Interview deleted: %1$s";
@@ -59,6 +59,9 @@ public class InterviewDeleteCommand extends InterviewCommand {
         }
 
         Job jobToDeleteInterview = lastShownList.get(jobIndex.getZeroBased());
+        if (interviewIndex.getZeroBased() >= jobToDeleteInterview.interviewLength()) {
+            throw new CommandException(Messages.MESSAGE_INVALID_INTERVIEW);
+        }
 
         jobToDeleteInterview.deleteInterview(interviewIndex);
         model.updateFilteredJobList(Model.PREDICATE_SHOW_ALL_JOBS);
